Created Mapper for Association rules
