+++ .gitignore
@@ -1,5 +1,7 @@
 bin/
 out/
 tmp/
+enhanced_in/
+outputRules/
 .project
 .classpath
